[Sprint 5] Feature user review ui #2

+++ app/src/main/java/com/framgia/fbook/data/model/Book.kt
@@ -72,7 +72,7 @@ class Book() : BaseModel(), Parcelable {
     var content: String? = null
     @Expose
     @SerializedName("star")
-    var star: Int? = null
+    var star: Float? = null
     @Expose
     @SerializedName("created_at")
     var createdAt: String? = null
